dom/indexedDB/test/test_setVersion_abort.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">

function testSteps()
{
  const READ_WRITE = Components.interfaces.nsIIDBTransaction.READ_WRITE;
  const VERSION_CHANGE = Components.interfaces.nsIIDBTransaction.VERSION_CHANGE;

  const name = window.location.pathname;
  const description = "My Test Database";

  let request = mozIndexedDB.open(name, 1, description);
  request.onerror = grabEventAndContinueHandler;
  request.onsuccess = errorHandler;
  request.onupgradeneeded = grabEventAndContinueHandler;
  let event = yield;

  let db = event.target.result;

  let objectStore = db.createObjectStore("foo");
  let index = objectStore.createIndex("bar", "baz");

  is(db.version, 1, "Correct version");
  is(db.objectStoreNames.length, 1, "Correct objectStoreNames length");
  is(objectStore.indexNames.length, 1, "Correct indexNames length");

  let transaction = event.target.transaction;
  transaction.oncomplete = unexpectedSuccessHandler;
  transaction.onabort = grabEventAndContinueHandler;
  transaction.abort();

  event = yield;
  is(event.type, "abort", "Got transaction abort event");
  is(event.target, transaction, "Right target");

  todo(db.version, 1, "Correct version");
  is(db.objectStoreNames.length, 0, "Correct objectStoreNames length");

  event = yield;
  is(event.type, "error", "Got request error event");
  is(event.target, request, "Right target");

  finishTest();
  yield;
}
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
